pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eless"
version = "1.0.0"
authors = [
    {name = "ELESS Team", email = "eless.team@example.com"},
]
description = "Evolving Low-resource Embedding and Storage System - A resilient RAG data processing pipeline with comprehensive logging, multi-database support, and CLI interface."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
    "Topic :: Database",
]
keywords = ["rag", "embedding", "vector-database", "nlp", "ai", "document-processing"]

# Core dependencies
dependencies = [
    "click>=8.0.0",
    "PyYAML>=6.0",
    "numpy>=1.21.0",
    "psutil>=5.8.0",
]

# Optional dependencies for different features
[project.optional-dependencies]
# Embedding models
embeddings = [
    "sentence-transformers>=2.2.0",
    "torch>=1.9.0",
]

# Vector databases
databases = [
    "chromadb>=0.4.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "qdrant-client>=1.6.0",
    "faiss-cpu>=1.7.0",
    "psycopg2-binary>=2.9.0",
    "cassandra-driver>=3.28.0",
]

# Document parsing
parsers = [
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "openpyxl>=3.1.0",
    "pandas>=1.5.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
]

# Full installation with all features
full = [
    "sentence-transformers>=2.2.0",
    "torch>=1.9.0",
    "chromadb>=0.4.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "qdrant-client>=1.6.0",
    "faiss-cpu>=1.7.0",
    "psycopg2-binary>=2.9.0",
    "cassandra-driver>=3.28.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "openpyxl>=3.1.0",
    "pandas>=1.5.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
]

[project.urls]
Homepage = "https://github.com/eless-team/eless"
Documentation = "https://github.com/eless-team/eless#readme"
Repository = "https://github.com/eless-team/eless.git"
"Bug Tracker" = "https://github.com/eless-team/eless/issues"
Changelog = "https://github.com/eless-team/eless/blob/main/CHANGELOG.md"

[project.scripts]
eless = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"src" = ["**/*.yaml", "**/*.yml", "**/*.json", "**/*.txt"]
"config" = ["*.yaml", "*.yml"]
"*" = ["*.yaml", "*.yml", "*.json", "*.txt"]

[tool.setuptools.package-dir]
"" = "."
config = "config"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true